requirements.txt
streamlit
numpy
pandas
scikit-learn
matplotlib
pillow
